+++ src/platform_impl/web/event_loop/runner.rs
@@ -1,5 +1,3 @@
-use super::proxy::{self, AsyncSender};
-use super::EventLoopProxy;
 use super::{super::ScaleChangeArgs, backend, state::State};
 use crate::event::{Event, StartCause};
 use crate::event_loop::ControlFlow;
@@ -29,12 +27,12 @@ pub struct Execution<T: 'static> {
     runner: RefCell<RunnerEnum<T>>,
     events: RefCell<VecDeque<Event<'static, T>>>,
     id: RefCell<u32>,
+    window: web_sys::Window,
     all_canvases: RefCell<Vec<(WindowId, Weak<RefCell<backend::Canvas>>)>>,
     redraw_pending: RefCell<HashSet<WindowId>>,
     destroy_pending: RefCell<VecDeque<WindowId>>,
     scale_change_detector: RefCell<Option<backend::ScaleChangeDetector>>,
     unload_event_handle: RefCell<Option<backend::UnloadEventHandle>>,
-    proxy_sender: AsyncSender<T>,
 }
 
 enum RunnerEnum<T: 'static> {
@@ -102,33 +100,22 @@ impl<T: 'static> Runner<T> {
 
 impl<T: 'static> Shared<T> {
     pub fn new() -> Self {
-        let (proxy_sender, mut proxy_receiver) = proxy::channel();
-
-        let this = Shared(Rc::new(Execution {
+        Shared(Rc::new(Execution {
             runner: RefCell::new(RunnerEnum::Pending),
             events: RefCell::new(VecDeque::new()),
+            #[allow(clippy::disallowed_methods)]
+            window: web_sys::window().expect("only callable from inside the `Window`"),
             id: RefCell::new(0),
             all_canvases: RefCell::new(Vec::new()),
             redraw_pending: RefCell::new(HashSet::new()),
             destroy_pending: RefCell::new(VecDeque::new()),
             scale_change_detector: RefCell::new(None),
             unload_event_handle: RefCell::new(None),
-            proxy_sender,
-        }));
+        }))
+    }
 
-        wasm_bindgen_futures::spawn_local({
-            let runner = this.clone();
-            async move {
-                while let Ok(value) = (&mut proxy_receiver).await {
-                    runner.send_event(Event::UserEvent(value))
-                }
-
-                // An error was returned because the channel was closed, which
-                // happens when the event loop gets closed, so we can stop now.
-            }
-        });
-
-        this
+    pub fn window(&self) -> &web_sys::Window {
+        &self.0.window
     }
 
     pub fn add_canvas(&self, id: WindowId, canvas: &Rc<RefCell<backend::Canvas>>) {
@@ -155,15 +142,19 @@ impl<T: 'static> Shared<T> {
 
         let close_instance = self.clone();
         *self.0.unload_event_handle.borrow_mut() =
-            Some(backend::on_unload(move || close_instance.handle_unload()));
+            Some(backend::on_unload(self.window(), move || {
+                close_instance.handle_unload()
+            }));
     }
 
     pub(crate) fn set_on_scale_change<F>(&self, handler: F)
     where
         F: 'static + FnMut(ScaleChangeArgs),
     {
-        *self.0.scale_change_detector.borrow_mut() =
-            Some(backend::ScaleChangeDetector::new(handler));
+        *self.0.scale_change_detector.borrow_mut() = Some(backend::ScaleChangeDetector::new(
+            self.window().clone(),
+            handler,
+        ));
     }
 
     // Generate a strictly increasing ID
@@ -173,10 +164,6 @@ impl<T: 'static> Shared<T> {
         *id += 1;
 
         *id
-    }
-
-    pub fn create_proxy(&self) -> EventLoopProxy<T> {
-        EventLoopProxy::new(self.0.proxy_sender.clone())
     }
 
     pub fn request_redraw(&self, id: WindowId) {
@@ -364,7 +351,7 @@ impl<T: 'static> Shared<T> {
             );
 
             // Then we resize the canvas to the new size and send a `Resized` event:
-            backend::set_canvas_size(&canvas, crate::dpi::Size::Physical(new_size));
+            backend::set_canvas_size(self.window(), &canvas, crate::dpi::Size::Physical(new_size));
             self.handle_single_event_sync(
                 Event::WindowEvent {
                     window_id: id,
@@ -455,7 +442,10 @@ impl<T: 'static> Shared<T> {
             ControlFlow::Poll => {
                 let cloned = self.clone();
                 State::Poll {
-                    request: backend::AnimationFrameRequest::new(move || cloned.poll()),
+                    request: backend::AnimationFrameRequest::new(
+                        self.window().clone(),
+                        move || cloned.poll(),
+                    ),
                 }
             }
             ControlFlow::Wait => State::Wait {
@@ -476,6 +466,7 @@ impl<T: 'static> Shared<T> {
                     start,
                     end,
                     timeout: backend::Timeout::new(
+                        self.window().clone(),
                         move || cloned.resume_time_reached(start, end),
                         delay,
                     ),