+++ src/platform_impl/web/web_sys/scaling.rs
@@ -8,29 +8,31 @@ use web_sys::MediaQueryListEvent;
 pub struct ScaleChangeDetector(Rc<RefCell<ScaleChangeDetectorInternal>>);
 
 impl ScaleChangeDetector {
-    pub(crate) fn new<F>(handler: F) -> Self
+    pub(crate) fn new<F>(window: web_sys::Window, handler: F) -> Self
     where
         F: 'static + FnMut(ScaleChangeArgs),
     {
-        Self(ScaleChangeDetectorInternal::new(handler))
+        Self(ScaleChangeDetectorInternal::new(window, handler))
     }
 }
 
 /// This is a helper type to help manage the `MediaQueryList` used for detecting
 /// changes of the `devicePixelRatio`.
 struct ScaleChangeDetectorInternal {
+    window: web_sys::Window,
     callback: Box<dyn FnMut(ScaleChangeArgs)>,
     mql: Option<MediaQueryListHandle>,
     last_scale: f64,
 }
 
 impl ScaleChangeDetectorInternal {
-    fn new<F>(handler: F) -> Rc<RefCell<Self>>
+    fn new<F>(window: web_sys::Window, handler: F) -> Rc<RefCell<Self>>
     where
         F: 'static + FnMut(ScaleChangeArgs),
     {
-        let current_scale = super::scale_factor();
+        let current_scale = super::scale_factor(&window);
         let new_self = Rc::new(RefCell::new(Self {
+            window: window.clone(),
             callback: Box::new(handler),
             mql: None,
             last_scale: current_scale,
@@ -43,7 +45,7 @@ impl ScaleChangeDetectorInternal {
             }
         }) as Box<dyn FnMut(_)>);
 
-        let mql = Self::create_mql(closure);
+        let mql = Self::create_mql(&window, closure);
         {
             let mut borrowed_self = new_self.borrow_mut();
             borrowed_self.mql = mql;
@@ -52,9 +54,10 @@ impl ScaleChangeDetectorInternal {
     }
 
     fn create_mql(
+        window: &web_sys::Window,
         closure: Closure<dyn FnMut(MediaQueryListEvent)>,
     ) -> Option<MediaQueryListHandle> {
-        let current_scale = super::scale_factor();
+        let current_scale = super::scale_factor(window);
         // This media query initially matches the current `devicePixelRatio`.
         // We add 0.0001 to the lower and upper bounds such that it won't fail
         // due to floating point precision limitations.
@@ -63,7 +66,7 @@ impl ScaleChangeDetectorInternal {
              (-webkit-min-device-pixel-ratio: {min_scale:.4}) and (-webkit-max-device-pixel-ratio: {max_scale:.4})",
             min_scale = current_scale - 0.0001, max_scale= current_scale + 0.0001,
         );
-        let mql = MediaQueryListHandle::new(&media_query, closure);
+        let mql = MediaQueryListHandle::new(window, &media_query, closure);
         if let Some(mql) = &mql {
             assert!(mql.mql().matches());
         }
@@ -76,12 +79,12 @@ impl ScaleChangeDetectorInternal {
             .take()
             .expect("DevicePixelRatioChangeDetector::mql should not be None");
         let closure = mql.remove();
-        let new_scale = super::scale_factor();
+        let new_scale = super::scale_factor(&self.window);
         (self.callback)(ScaleChangeArgs {
             old_scale: self.last_scale,
             new_scale,
         });
-        let new_mql = Self::create_mql(closure);
+        let new_mql = Self::create_mql(&self.window, closure);
         self.mql = new_mql;
         self.last_scale = new_scale;
     }