+++ tests/send_objects.rs
@@ -10,7 +10,6 @@ fn event_loop_proxy_send() {
     }
 }
 
-#[cfg(not(wasm_platform))]
 #[test]
 fn window_send() {
     // ensures that `winit::Window` implements `Send`